requirements.txt
amqp==2.5.2
billiard==3.6.3.0
Django==2.2
# django-cors-headers==3.2.1
django-filter==2.2.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
djangorestframework==3.10.3
drf-renderer-xlsx==0.3.7
et-xmlfile==1.0.1
importlib-metadata==1.5.0
jdcal==1.4.1
jsonfield==3.1.0
kombu==4.6.8
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.3
redis==3.4.1
sqlparse==0.3.0
swapper==1.1.2.post1
vine==1.3.0
xlrd==1.2.0
zipp==3.1.0
